pyproject.toml
[tool.poetry]
name = "fastapi-app"
version = "0.1.0"
description = ""
authors = ["cnwv <115669384+cnwv@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
alembic = "^1.13.2"
fastapi-users = { extras = ["sqlalchemy"], version = "^13.0.0" }
sqlalchemy = { extras = ["asyncio"], version = "^2.0.31" }
asyncpg = "^0.29.0"
uvicorn = { extras = ["standart"], version = "^0.30.1" }
pydantic-settings = "^2.3.4"
aiobotocore = "^2.13.1"
pyjwt = { extras = ["crypto"], version = "^2.8.0" }
python-multipart = "^0.0.9"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
black = "^24.4.2"
pre-commit = "^3.8.0"
mypy = "^1.11.1"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | foo.py           # also separately exclude a file named foo.py in
                     # the root of the project
)
'''
[tool.flake8]
max-line-length = 88

[tool.mypy]
ignore_missing_imports = true
